+++ docs/src/content/functions/content.config.ts
@@ -1,36 +1,17 @@
 import { defineCollection, reference, z } from "astro:content";
-import { map, pick } from "remeda";
-import invariant from "tiny-invariant";
-import { Application as TypeDoc, type ProjectReflection } from "typedoc";
 import { zEntry } from "./schema";
+import { categoriesLoader, functionsLoader } from "./loaders";
 
 export const functionsCollectionName = "functions";
 export const categoriesCollectionName = "categories";
 
 export const functionsCollection = defineCollection({
-  loader: async () =>
-    await fromTypeDoc(
-      "children",
-      map(({ name, ...rest }) => ({
-        id: name,
-        name,
-        ...pick(rest, ["kind", "signatures", "sources"]),
-      })),
-    ),
-
+  loader: functionsLoader,
   schema: zEntry,
 });
 
 export const categoriesCollection = defineCollection({
-  loader: async () =>
-    await fromTypeDoc(
-      "categories",
-      map(({ title: id, children }) => ({
-        id,
-        children: children.map(({ name }) => name),
-      })),
-    ),
-
+  loader: categoriesLoader,
   schema: z
     .object({
       id: z.string(),
@@ -38,18 +19,3 @@ export const categoriesCollection = defineCollection({
     })
     .strict(),
 });
-
-async function fromTypeDoc<K extends keyof ProjectReflection, T>(
-  key: K,
-  extractor: (data: NonNullable<ProjectReflection[K]>) => T,
-): Promise<T> {
-  const app = await TypeDoc.bootstrapWithPlugins();
-
-  const project = await app.convert();
-  invariant(project !== undefined, "Failed to parse project!");
-
-  const val = project[key];
-  invariant(val !== undefined, `Parsed project is missing '${key}' data`);
-
-  return extractor(val);
-}
