+++ docs/src/content/functions/v1/content.config.ts
@@ -1,21 +1,17 @@
 import { defineCollection, reference, z } from "astro:content";
-import path from "node:path";
 import { zEntry } from "../schema";
-import dataFilePath from "./functions.json?url";
 import { categoriesLoader, functionsLoader } from "./loaders";
-
-const DATA_FILE = path.join(import.meta.dirname, path.basename(dataFilePath));
 
 export const functionsV1CollectionName = "functions-v1";
 export const categoriesV1CollectionName = "categories-v1";
 
 export const functionsV1Collection = defineCollection({
-  loader: functionsLoader(DATA_FILE),
+  loader: functionsLoader,
   schema: zEntry,
 });
 
 export const categoriesV1Collection = defineCollection({
-  loader: categoriesLoader(DATA_FILE),
+  loader: categoriesLoader,
   schema: z
     .object({
       id: z.string(),
